Initial draft of checking out with fallback branch
Summary
=======

Adds a composite action file that can be included in other GitHub Actions, along with a README about usage.

Motivation
----------

My intent is that this can be used to checkout out specific branches, such as when tests require matching branches across repos. See README for more specifics.

Risks
-----

I don't know how to test a GitHub Action without running it, so this code has not been tested. I'll iterate on it in subsequent commits if it does not work where I try to include it.

Testing
=======

None, but I will test it separately in another repo's workflow.

Issues
======

N/A